// src/domain/enums/Finger.ts
/**
 * Finger enumeration for keyboard mapping
 */
export enum Finger {
  PINKY_LEFT = 'pinky-left',
  RING_LEFT = 'ring-left',
  MIDDLE_LEFT = 'middle-left',
  INDEX_LEFT = 'index-left',
  THUMB = 'thumb',
  INDEX_RIGHT = 'index-right',
  MIDDLE_RIGHT = 'middle-right',
  RING_RIGHT = 'ring-right',
  PINKY_RIGHT = 'pinky-right',
}